Stock Price Prediction/requirements.txt
matplotlib==3.7.2
numpy==1.25.1
pandas==2.0.3
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
yfinance==0.2.24